# backend/storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'E-commerce Modular API'
  description: 'Tài liệu API hệ thống E-commerce (Sử dụng PHP 8 Attributes)'
  contact:
    email: admin@system.com
  version: 1.0.0
servers:
  -
    url: /api
    description: 'API Server'
paths:
...
  /admin/inventories:
    get:
      tags:
        - Inventory
      summary: 'Xem danh sách tồn kho'
      operationId: 60f3767a8279f38400634bcec6ff93a0
      parameters:
        -
          name: page
          in: query
          schema:
            type: integer
        -
          name: per_page
          in: query
          schema:
            type: integer
        -
          name: product_uuid
          in: query
          description: 'Filter theo UUID sản phẩm'
          required: false
          schema:
            type: string
            format: uuid
        -
          name: warehouse_uuid
          in: query
          description: 'Filter theo UUID kho'
          required: false
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: 'Successful operation'
        403:
          description: 'Forbidden (Not Admin)'
      security:
        -
          bearerAuth: []
  /admin/inventories/upsert:
    post:
      tags:
        - Inventory
      summary: 'Thêm hoặc Cập nhật Tồn kho (Upsert)'
      description: 'Tạo record tồn kho mới hoặc ghi đè (overwrite) số lượng nếu đã tồn tại.'
      operationId: 109e5267df7045ba5112570cb42f07a3
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - product_uuid
                - warehouse_uuid
                - quantity
              properties:
                product_uuid:
                  type: string
                  format: uuid
                warehouse_uuid:
                  type: string
                  format: uuid
                quantity:
                  description: 'Số lượng tồn kho mới (overwrite)'
                  type: integer
                min_threshold:
                  description: 'Ngưỡng cảnh báo thấp nhất'
                  type: integer
                  nullable: true
              type: object
      responses:
        200:
          description: 'Inventory updated successfully'
        403:
          description: Forbidden
      security:
        -
          bearerAuth: []
  '/admin/inventories/{productUuid}/{warehouseUuid}/adjust':
    patch:
      tags:
        - Inventory
      summary: 'Điều chỉnh số lượng tồn kho (Tăng/Giảm Delta)'
      description: 'Sử dụng để nhập kho (delta > 0) hoặc trừ kho thủ công (delta < 0).'
      operationId: b649baf17cf0e8e3e443180f3cd41952
      parameters:
        -
          name: productUuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
        -
          name: warehouseUuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - delta
              properties:
                delta:
                  description: 'Giá trị thay đổi (có thể âm hoặc dương)'
                  type: integer
              type: object
      responses:
        200:
          description: 'Stock adjusted successfully'
        403:
          description: Forbidden
        422:
          description: 'Insufficient stock or Validation error'
      security:
        -
          bearerAuth: []
  /admin/logs:
    get:
      tags:
        - 'System Logs'
      summary: 'Xem danh sách Log'
      operationId: 682c7487c29b68fd9dbbd61477696034
      parameters:
        -
          name: page
          in: query
          schema:
            type: integer
        -
          name: per_page
          in: query
          schema:
            type: integer
        -
          name: level
          in: query
          description: 'error, info, warning...'
          schema:
            type: string
      responses:
        200:
          description: Success
        403:
          description: Forbidden
      security:
        -
          bearerAuth: []
  '/admin/logs/{uuid}':
    get:
      tags:
        - 'System Logs'
      summary: 'Xem chi tiết Log'
      operationId: c7c9b88cabffbd44d6ef387e741c7f4d
      parameters:
        -
          name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Success
      security:
        -
          bearerAuth: []
...
  /orders:
    get:
      tags:
        - Orders
      summary: 'Xem danh sách đơn hàng'
      operationId: 4e011b1dd08685b7386fdce144bf8f7b
      parameters:
        -
          name: status
          in: query
          schema:
            type: string
        -
          name: page
          in: query
          schema:
            type: integer
        -
          name: per_page
          in: query
          schema:
            type: integer
      responses:
        200:
          description: Success
      security:
        -
          bearerAuth: []
    post:
      tags:
        - Orders
      summary: 'Tạo đơn hàng mới (Admin/Staff)'
      operationId: 6dc386d376574a4249e0ddb0cbcb2359
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - address_id
                - items
              properties:
                address_id:
                  type: integer
                items:
                  type: array
                  items:
                    type: object
              type: object
      responses:
        201:
          description: Created
      security:
        -
          bearerAuth: []
  /orders/checkout:
    post:
      tags:
        - Orders
      summary: 'Đặt hàng từ giỏ hàng (User)'
      operationId: d214f391766889d74685377a5f161aaa
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - address_id
              properties:
                address_id:
                  type: integer
                notes:
                  type: string
              type: object
      responses:
        201:
          description: 'Order placed'
      security:
        -
          bearerAuth: []
  '/orders/{uuid}':
    get:
      tags:
        - Orders
      summary: 'Xem chi tiết đơn hàng'
      operationId: f15bfcfbe0db066bb7ed1468987c7689
      parameters:
        -
          name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Success
        403:
          description: Forbidden
      security:
        -
          bearerAuth: []
  '/orders/{uuid}/cancel':
    post:
      tags:
        - Orders
      summary: 'Hủy đơn hàng'
      operationId: 68403667692512eb7b881131b53ff0d5
      parameters:
        -
          name: uuid
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Cancelled
      security:
        -
          bearerAuth: []
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  -
    bearerAuth: []